feat: consolidate Weblate CLI; move into core
