+++ Core/Sources/Weblate/Foundation/JSONDecoder.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+extension JSONDecoder {
+    convenience init(_ keyDecodingStrategy: KeyDecodingStrategy) {
+        self.init()
+        self.keyDecodingStrategy = keyDecodingStrategy
+    }
+}
